pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.67.0",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
]
